--- src/com/company/Menus/Login.java
@@ -1,19 +1,0 @@
-package com.company.Menus;
-
-import javax.swing.*;
-
-public class Login {
-    private JPanel panel1;
-    private JTextField textField1;
-    private JPasswordField passwordField1;
-
-    public void setData(Test data) {
-    }
-
-    public void getData(Test data) {
-    }
-
-    public boolean isModified(Test data) {
-        return false;
-    }
-}
