--- src/com/company/Menus/Test.java
@@ -1,6 +1,0 @@
-package com.company.Menus;
-
-public class Test {
-    public Test() {
-    }
-}
